pyproject.toml
[tool.poetry]
name = "rest-pg"
version = "0.1.0"
description = ""
authors = ["Mihail Butvin <butvin.mihail@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
pydantic = "^2.5.2"
python-dotenv = "^1.0.0"
psycopg = {extras = ["binary"], version = "^3.1.14"}

[tool.poetry.group.dev.dependencies]
mypy = "^1.7.1"
isort = "^5.12.0"
wemake-python-styleguide = "^0.18.0"
pytest = "^7.4.1"
pytest-asyncio = "^0.21.1"
testcontainers-postgres = "^0.0.1rc1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
